// src/api/openapi/models/offerCreate.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DescriptiveArticleCreate } from './descriptiveArticleCreate';


export interface OfferCreate { 
    date: string;
    delivery: string;
    descriptive_articles: Array<DescriptiveArticleCreate>;
    discount_amount: number;
    discount_percentage: number;
    in_price_included: string;
    job_id: number;
    material_description: string;
    material_description_title: string;
    payment: string;
    validity: string;
    vat_id: number;
}