// src/api/openapi/model/descriptiveArticle.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Vat } from './vat';
import { Unit } from './unit';


export interface DescriptiveArticle { 
    alternative: boolean;
    amount: number;
    description: string;
    descriptive_article?: Array<DescriptiveArticle>;
    discount: number;
    header?: boolean;
    id: number;
    name: string;
    price_with_vat?: number;
    price_without_vat?: number;
    single_price: number;
    unit: Unit;
    vat: Vat;
    vat_amount?: number;
}